refactor: extrai trilhas e a checagem de admin para services
Move o CRUD e a leitura de trilhas (criar, atualizar, excluir, listar, minhas trilhas e o detalhe com o estado das aulas) para trail.service. O helper ehAdmin vai para usuario.service, reusado pelo detalhe da trilha e ainda pelas aulas e quiz no controller. Comportamento da API preservado.
